Create README.md
PØØR is a token developed by The Philosopher inside DistrictOne app on Blast Chain.
Its idea was to reunite people that identify themselves with the concept and elevate 
their power through community projects and funding. We decided that communities are 
never poor, and their abilities, when put together with communal intentions, can bring 
incredible things into existence. If you identify with the concept, please come with us 
because we are building beautiful things together. When people with the same intention 
are together, there is nothing they cant achieve. 

One of the first things we are trying to build together is winning the mini-eth jackpot 
from DistrictOne. This is only to be won by the top performing coins. So, as our first 
communal action, we are trying to develop our coin the best we can, in number of 24h 
swaps, market cap, price, active users, holders, and many other criteria. If we get to 
be one of the winners, all holders will earn eth according to their holdings. This will 
be a beautiful first achievement.

We also created some NFTs to support our community and treasury. Some NFTs were listed 
in PØØR tokens, in order to stimulate the use of the token and make it go beyond a mere 
meme token. Check our NFT page, to support our project. All the resources coming from NFT 
sales are used to buy PØØR and increase its market cap. All periodically free NFTs are 
distributed to our holders.

If you want to know more about our token, please check this site and the info below:

- PØØR D1 Community
- PØØR Community DAO on Snapshot
- Buy PØØR (Thruster)
- PØØR on Dextools
- PØØR Contract Address: 0xb392684843e9edf7adf9e30bb0475bf44fbcd5a9 (Blast Chain)

Check our blog/site for more info: https://poor-community.blogspot.com/